// Games/src/app/models/heromodel.ts
export class heromodel{
   HeroId: number;
   HeroName: string;
   HeroAttackMin: number;
   HeroAttackMax: number;
   HeroDefenceMin: number;
   HeroDefenceMax: number;
   HeroPotMin: number;
   HeroPotMax: number;
   HeroHealth: number;
   HeroLevel: number;
   MaxHealth: number;
   HeroImagePath: string | null;
}